requirements.txt
alembic==1.4.3
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.1
asn1crypto==0.24.0
astroid==2.4.2
attrs==20.2.0
auth==0.5.3
autopep8==1.5.4
beautifulsoup4==4.9.3
blinker==1.4
bootstrap4==0.1.0
Brlapi==0.6.6
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cloudinary==1.23.0
command-not-found==0.3
config==0.5.0.post0
confusable-homoglyphs==3.2.0
contextvars==2.4
coverage==5.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
devconfig==0.4.8
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap3==15.0.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-core==1.4.1
django-cors-headers==3.5.0
django-crispy-forms==1.11.2
django-decouple==2.1
django-heroku==0.3.1
django-model-utils==4.0.0
django-registration==3.1.1
django-sendgrid-v5==0.9.1
django-star-ratings==0.9.1
django-tinymce==3.1.0
django-vote==2.2.0
djangorestframework==3.12.2
dnspython==1.16.0
dominate==2.6.0
email-validator==1.1.1
eventlet==0.29.1
falcon==2.0.0
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
future==0.18.2
greencache==1.0.0
greenlet==0.4.17
gunicorn==20.0.4
htmq==1.0.0
httplib2==0.9.2
idna==2.10
immutables==0.14
importlib-metadata==2.1.0
importlib-resources==3.0.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
markdown2==2.3.10
MarkupSafe==1.0
mccabe==0.6.1
mongoengine==0.20.0
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==8.0.1
protobuf==3.0.0
psycopg2==2.8.6
pycairo==1.16.2
pycodestyle==2.6.0
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.6.0
pymacaroons==0.13.0
pymongo==3.11.0
PyNaCl==1.1.2
pyperclip==1.8.1
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
python-decouple==3.3
python-editor==1.0.4
python-http-client==3.3.2
python-json-logger==2.0.1
pytz==2020.4
pyuploadcare==2.7.0
pyxdg==0.25
PyYAML==3.12
redcmd==1.2.10
redlib==1.5.7
reportlab==3.4.0
requests==2.25.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
sendgrid==6.7.0
simplejson==3.13.2
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
swapper==1.1.2.post1
system-service==0.3
systemd-python==234
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.10.0.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.26.2
usb-creator==0.3.3
virtualenv==20.0.34
visitor==0.1.3
vote==1.0.2
votes==1.0.1
voting==0.1.3
wadllib==1.3.2
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
WTForms==2.3.3
xkit==0.0.0
zipp==3.4.0
zope.interface==5.2.0
